Changed Movie Object structure & Added Genre class
Changed some of the elements in Movie.js to match the names of the moviedb api. Genres in the moviedb api are represented as objects and not simply titles.
Object is made of the id and name. Created a separate class for Genres and made a list of genres as a component for Movie